Add support for 64-bit operation status values
For compatibility with 64-bit EFI status codes, opstatus values
are now treated as 64-bit rather than 32-bit. A new rpc_opstatus_t
type has been introduced in this change. Note that the FFARPC
still maps the opstatus value to a 32-bit regsiter value but
this is sign extended up to 64-bits in RPC callers.  The MM
Communicate RPC caller and endpoint handle 64-bit status values.

+++ protocols/common/efi/efi_status.h
@@ -13,10 +13,10 @@
 /**
  * Common EFI status type
  */
-typedef uint32_t efi_status_t;
+typedef uint64_t efi_status_t;
 
 /* For error class status codes */
-#define EFI_ERROR(s)				((1U << ((CHAR_BIT * sizeof(efi_status_t)) - 1)) | s)
+#define EFI_ERROR(s)	((1ULL << ((CHAR_BIT * sizeof(efi_status_t)) - 1)) | s)
 
 /**
  * Common EFI status codes
